don't import nncf in conftest to not fail install tests
